Improve keyboard shortcuts for yiddish and update documentation
A function was added to the keyboardShortcuts.js that toggles notifications to align with the current code style. Keyboard shortcuts were updated to work when the Yiddish keyboard is on. The documentation in settings.html and README.md was updated to add information on new and updated keyboard shortcuts, thereby providing clear user guidance.