Update extension to version 2.1.2 with script improvements
This commit updates the version of the extension from 2.1.1 to 2.1.2 in manifest.json. It also brings changes in the addTopicButtons.js script where unnecessary attribute 'itext' from the 'createButton' and 'addSimpleButton' functions were removed. An equality check was added to correctly activate a conditional statement in yet another function. Typos in notification messages have been corrected as well.