Add icon text property to createButton function
The createButton function and its uses have been expanded to accept an additional parameter, which is used to set the text of the button's icon. This parameter is optional and has a default value of empty string. Several hardcoded text values, and CSS style adjustments have been made for the icon text.