Added a new 'print' functionality to the extension
This update introduces a new printing feature to the extension that enhances the print layout of thread posts. This function, derived from bookmarkify, has been integrated into the settings of the extension and can be toggled on or off based on user preference.Update bookmarkify reference link in print.js

The bookmarkify referenced link at the start of the print.js script has been updated. The previously linked page was replaced with a new, more recent one, providing up-to-date methods and functions.